+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - improved the view of debugger stack frames
 - show source for system libraries when debugging
 - handle breakpoints in libraries loaded as self-references
+- have this plugin depend on the `flutter` plugin
 
 ## 0.5.0
 - we now show errors in the outline view